refactor: upgrade dependencies

+++ frontend/src/page/Playground.tsx
@@ -1,24 +1,24 @@
 import infoList from "../list";
 import "./PlayGround.css";
-import useStyles from "../util/style";
 
 import { apiUrl } from "../App";
-import { previousContent, nextContent } from "../util/contentHandler";
+import { nextContent, previousContent } from "../util/contentHandler";
 
 /* eslint-disable-next-line import/no-webpack-loader-syntax */
 import Worker from "worker-loader!../util/runtime.worker";
 
-import { Redirect, Link, useParams } from "react-router-dom";
+import { Link, Navigate, useNavigate, useParams } from "react-router-dom";
 import {
   Alert,
+  AlertColor,
   Box,
   Button,
   Grid,
   Paper,
   Snackbar,
   Typography,
-} from "@material-ui/core";
-import React, { useEffect, useRef, useState } from "react";
+} from "@mui/material";
+import React, { Fragment, useEffect, useRef, useState } from "react";
 
 import AceEditor from "react-ace";
 import "ace-builds/webpack-resolver";
@@ -34,14 +34,13 @@ import ansiColor from "ansi-colors";
 
 import Markdown from "markdown-to-jsx";
 import axios from "axios";
-import { Color as AlertColor } from "@material-ui/core/Alert";
 
-import DialogTitle from "@material-ui/core/DialogTitle";
-import DialogContent from "@material-ui/core/DialogContent";
-import DialogActions from "@material-ui/core/DialogActions";
-import Dialog from "@material-ui/core/Dialog";
+import DialogTitle from "@mui/material/DialogTitle";
+import DialogContent from "@mui/material/DialogContent";
+import DialogActions from "@mui/material/DialogActions";
+import Dialog from "@mui/material/Dialog";
 
-interface PlaygroundRouteParams {
+interface PlaygroundRouteParams extends Record<string, string | undefined> {
   section: string;
   content: string;
 }
@@ -49,7 +48,7 @@ interface PlaygroundRouteParams {
 function Playground(): JSX.Element {
   const { section, content } = useParams<PlaygroundRouteParams>();
 
-  const classes = useStyles();
+  const navigate = useNavigate();
 
   const [code, setCode] = useState("");
   const [article, setArticle] = useState("");
@@ -70,9 +69,7 @@ function Playground(): JSX.Element {
     msg: "",
   });
 
-  const terminal = useRef(
-    new Terminal({ cursorBlink: true, cursorStyle: "bar" })
-  );
+  const terminal = useRef<Terminal>();
 
   const workerRef = useRef<Worker>();
 
@@ -88,7 +85,7 @@ function Playground(): JSX.Element {
   const runCode = (e: React.MouseEvent<HTMLButtonElement>) => {
     e.preventDefault();
 
-    terminal.current.reset();
+    terminal.current?.reset();
     setLoading(true);
 
     axios
@@ -109,7 +106,7 @@ function Playground(): JSX.Element {
           let errors = (resp.data.error as string).split("\n");
           for (let i = 0; i < errors.length - 2; i++) {
             showErrors.push(
-              <div className={classes.codeFont} key={i}>
+              <div className={"codeFont"} key={i}>
                 {errors[i]
                   .replaceAll(" ", "\u00A0")
                   .replace(/\/tmp\/\d+.c/, "/tmp/code.c")}
@@ -138,32 +135,34 @@ function Playground(): JSX.Element {
   };
 
   useEffect(() => {
-    import(`!!raw-loader!../document/${section}/${content}.md`).then(
-      (article) => {
-        setArticle(article.default);
-      }
-    );
+    const fetchContent = async (url: string, setter: React.Dispatch<any>) => {
+      const value = await fetch(url).then((resp) => resp.text());
+      setter(value);
+    };
+
+    if (section === undefined || content === undefined) {
+      navigate("/404");
+      return;
+    }
+
+    import(`../document/${section}/${content}.md`).then((article) => {
+      fetchContent(article.default, setArticle);
+    });
     if (infoList[section].content[content].hasCode) {
-      import(`!!raw-loader!../code/${section}/${content}.c`).then((code) => {
-        setCode(code.default);
+      import(`../code/${section}/${content}.c`).then((code) => {
+        fetchContent(code.default, setCode);
       });
     } else {
       setCode("");
     }
 
-    terminal.current.reset();
-  }, [content, section]);
+    terminal.current?.reset();
+  }, [content, section, navigate]);
 
   useEffect(() => {
-    const term = terminal.current;
-    const fitAddon = new FitAddon();
-    term.loadAddon(fitAddon);
-    term.open(document.getElementById("terminal") as HTMLElement);
-    fitAddon.fit();
-
     const worker = new Worker();
 
-    const mem = new SharedArrayBuffer(4104);
+    const mem = new SharedArrayBuffer(20 * 1024 * 1024);
     const stdinBuffer = new Int32Array(mem);
     let stdinPtr = 0;
 
@@ -177,44 +176,52 @@ function Playground(): JSX.Element {
       Atomics.notify(stdinBuffer, stdinBuffer.length - 1);
     };
 
-    term.onData((e) => {
-      switch (e) {
-        case "\r": {
-          term.write("\r\n");
-          stdinBuffer[stdinPtr] = e.charCodeAt(0);
-          stdinPtr++;
-          Atomics.store(stdinBuffer, stdinBuffer.length - 1, stdinPtr);
-          wakeupWorker();
-          break;
-        }
-        case "\u0003": // Ctrl+C
-          break;
-        case "\u007F": // Backspace (DEL)
-          // @ts-ignore
-          if (term._core.buffer.x > 0) {
-            term.write("\b \b");
-            // // @ts-ignore
-            // console.log(term._core);
-            stdinBuffer[stdinPtr] = 0;
-            stdinPtr--;
+    if (terminal.current === undefined) {
+      const term = new Terminal({ cursorBlink: true, cursorStyle: "bar" });
+      const fitAddon = new FitAddon();
+      term.loadAddon(fitAddon);
+      term.open(document.getElementById("terminal") as HTMLElement);
+      fitAddon.fit();
+      console.log(fitAddon);
+      terminal.current = term;
+
+      term.onData((e) => {
+        switch (e) {
+          case "\r": {
+            term.write("\r\n");
+            stdinBuffer[stdinPtr] = e.charCodeAt(0);
+            stdinPtr++;
             Atomics.store(stdinBuffer, stdinBuffer.length - 1, stdinPtr);
-            worker.postMessage({
-              type: "delete",
-            });
+            wakeupWorker();
+            break;
           }
-          break;
-        default:
-          term.write(e);
-          stdinBuffer[stdinPtr] = e.charCodeAt(0);
-          console.log(stdinBuffer);
-          stdinPtr++;
-      }
-    });
+          case "\u0003": // Ctrl+C
+            break;
+          case "\u007F": // Backspace (DEL)
+            // @ts-ignore
+            if (term._core.buffer.x > 0) {
+              term.write("\b \b");
+              stdinBuffer[stdinPtr] = 0;
+              stdinPtr--;
+              Atomics.store(stdinBuffer, stdinBuffer.length - 1, stdinPtr);
+              worker.postMessage({
+                type: "delete",
+              });
+            }
+            break;
+          default:
+            term.write(e);
+            stdinBuffer[stdinPtr] = e.charCodeAt(0);
+            //console.log(stdinBuffer);
+            stdinPtr++;
+        }
+      });
+    }
 
     let rflag = false;
     const stdout = (keyCode: number) => {
       if (keyCode === null) {
-        terminal.current.reset();
+        terminal.current?.reset();
       }
 
       let str = String.fromCharCode(keyCode);
@@ -234,7 +241,7 @@ function Playground(): JSX.Element {
           rflag = false;
         }
       }
-      terminal.current.write(str);
+      terminal.current?.write(str);
     };
 
     worker.onmessage = (ev) => {
@@ -246,11 +253,15 @@ function Playground(): JSX.Element {
         }
         case "exit": {
           setLoading(false);
-          term.writeln("");
+          terminal.current?.writeln("");
           if (msg.data === 0) {
-            term.writeln(ansiColor.green.bold(`Exit with ${msg.data}`));
+            terminal.current?.writeln(
+              ansiColor.green.bold(`Exit with ${msg.data}`)
+            );
           } else {
-            term.writeln(ansiColor.red.bold(`Exit with ${msg.data}`));
+            terminal.current?.writeln(
+              ansiColor.red.bold(`Exit with ${msg.data}`)
+            );
           }
         }
       }
@@ -267,12 +278,17 @@ function Playground(): JSX.Element {
     setSnackbar({ status: false });
   };
 
-  if (!(section in infoList) || !(content in infoList[section].content)) {
-    return <Redirect to="/404" />;
+  if (
+    !section ||
+    !content ||
+    !(section in infoList) ||
+    !(content in infoList[section].content)
+  ) {
+    return <Navigate to="/404" />;
   }
 
   return (
-    <Box className={classes.container}>
+    <Fragment>
       <Dialog
         open={dialog.status}
         onClose={() => setDialog({ status: false })}
@@ -298,126 +314,150 @@ function Playground(): JSX.Element {
       <Snackbar
         open={snackbar.status}
         autoHideDuration={6000}
-        onClose={handleClose}
         anchorOrigin={{ vertical: "top", horizontal: "center" }}
       >
         <Alert onClose={handleClose} severity={snackbar.type}>
           {snackbar.msg}
         </Alert>
       </Snackbar>
-      <Grid container spacing={2} className={classes.divideBox}>
-        <Grid item xs={6} className={classes.divideBox}>
-          <Box className={classes.cardBox0}>
-            <Paper
-              className={[classes.paper, classes.scroll].join(" ")}
-              elevation={2}
-              style={{ padding: "16px" }}
-            >
-              <Typography variant={"h2"} fontWeight={"bold"} fontSize={"24px"}>
-                {infoList[section].content[content].title}
-              </Typography>
-              <Box style={{ marginBottom: "56px" }}>
-                <Markdown options={{ wrapper: "article" }}>{article}</Markdown>
-              </Box>
-            </Paper>
-          </Box>
-          <Box className={classes.buttonBar}>
-            <Button
-              variant="outlined"
-              color="primary"
-              disableElevation
-              component={Link}
-              to={getPath(previousContent)}
-            >
-              上一页
-            </Button>
-            <Button
-              variant="outlined"
-              color="primary"
-              disableElevation
-              component={Link}
-              to="/catalogue"
-            >
-              目录
-            </Button>
-            <Button
-              variant="outlined"
-              color="primary"
-              disableElevation
-              component={Link}
-              to={getPath(nextContent)}
-            >
-              下一页
-            </Button>
-          </Box>
-        </Grid>
-        <Grid item xs={6} className={classes.divideBox}>
-          <Box className={classes.cardBox1}>
-            <Paper
-              className={classes.paper}
-              elevation={2}
-              style={{ borderRadius: 0 }}
-            >
-              <AceEditor
-                style={{
-                  width: "100%",
-                  height: "100%",
-                  fontFamily: "'Consolas', sans-serif",
-                }}
-                mode="c_cpp"
-                theme="github"
-                name="editor"
-                fontSize={16}
-                showPrintMargin={true}
-                showGutter={true}
-                highlightActiveLine={true}
-                value={code}
-                debounceChangePeriod={200}
-                setOptions={{
-                  enableBasicAutocompletion: true,
-                  enableLiveAutocompletion: true,
-                  enableSnippets: true,
-                  showLineNumbers: true,
-                  tabSize: 2,
-                }}
-                onChange={onEditorChange}
-              />
-            </Paper>
-            <Box className={classes.submitButton}>
-              <Button
-                variant="contained"
-                color="primary"
-                style={{
-                  float: "right",
-                  marginRight: "16px",
-                }}
-                disabled={loading}
-                onClick={runCode}
+      <main
+        style={{
+          height: "calc(100vh - 48px)",
+        }}
+      >
+        <Grid
+          container
+          spacing={2}
+          style={{
+            height: "100%",
+          }}
+        >
+          <Grid
+            item
+            xs={6}
+            style={{
+              maxHeight: "100%",
+              paddingTop: "0",
+            }}
+          >
+            <Box className={"cardBox0 scroll"}>
+              <Paper
+                className={"paper scroll"}
+                elevation={2}
+                style={{ padding: "16px" }}
               >
-                运行
+                <Typography
+                  variant={"h2"}
+                  fontWeight={"bold"}
+                  fontSize={"24px"}
+                >
+                  {infoList[section].content[content].title}
+                </Typography>
+                <Box style={{ marginBottom: "56px" }}>
+                  <Markdown options={{ wrapper: "article" }}>
+                    {article}
+                  </Markdown>
+                </Box>
+              </Paper>
+            </Box>
+            <Box className={"buttonBar"}>
+              <Button
+                variant="outlined"
+                color="primary"
+                disableElevation
+                component={Link}
+                to={getPath(previousContent)}
+              >
+                上一页
+              </Button>
+              <Button
+                variant="outlined"
+                color="primary"
+                disableElevation
+                component={Link}
+                to="/catalogue"
+              >
+                目录
+              </Button>
+              <Button
+                variant="outlined"
+                color="primary"
+                disableElevation
+                component={Link}
+                to={getPath(nextContent)}
+              >
+                下一页
               </Button>
             </Box>
-          </Box>
-          <Box className={classes.cardBox2}>
-            <Paper
-              className={classes.paper}
-              elevation={2}
-              style={{ borderRadius: 0 }}
-            >
-              <Box
-                id="terminal"
-                style={{
-                  height: "100%",
-                  width: "100%",
-                  padding: "8px",
-                  backgroundColor: "rgb(0, 0, 0)",
-                }}
-              />
-            </Paper>
-          </Box>
+          </Grid>
+          <Grid item xs={6} style={{ padding: 0 }}>
+            <Box className={"cardBox1"}>
+              <Paper
+                className={"paper"}
+                elevation={2}
+                style={{ borderRadius: 0 }}
+              >
+                <AceEditor
+                  style={{
+                    width: "100%",
+                    height: "100%",
+                    fontFamily: "'Consolas', sans-serif",
+                  }}
+                  mode="c_cpp"
+                  theme="github"
+                  name="editor"
+                  fontSize={16}
+                  showPrintMargin={true}
+                  showGutter={true}
+                  highlightActiveLine={true}
+                  value={code}
+                  debounceChangePeriod={200}
+                  setOptions={{
+                    enableBasicAutocompletion: true,
+                    enableLiveAutocompletion: true,
+                    enableSnippets: true,
+                    showLineNumbers: true,
+                    tabSize: 2,
+                  }}
+                  onChange={onEditorChange}
+                />
+              </Paper>
+              <Box className={"submitButton"}>
+                <Button
+                  variant="contained"
+                  color="primary"
+                  style={{
+                    float: "right",
+                    marginRight: "16px",
+                  }}
+                  disabled={loading}
+                  onClick={runCode}
+                >
+                  运行
+                </Button>
+              </Box>
+            </Box>
+            <Box className={"cardBox2"}>
+              <Paper
+                className={"paper"}
+                elevation={2}
+                style={{ borderRadius: 0 }}
+              >
+                <Box
+                  id="terminal"
+                  style={{
+                    height: "100%",
+                    width: "100%",
+                    padding: "8px",
+                    backgroundColor: "rgb(0, 0, 0)",
+                  }}
+                />
+              </Paper>
+            </Box>
+          </Grid>
         </Grid>
-      </Grid>
-    </Box>
+      </main>
+    </Fragment>
   );
 }
